fix: 20个模型BigInteger→Integer (SQLite兼容) + Integer/BigInteger双导入
- 所有模型PK从BigInteger改为Integer(SQLite autoincrement支持)
- 非PK字段保留BigInteger
- 测试: 31→71 passed (+40修复)

+++ backend/app/models/acceptance.py
@@ -1,10 +1,10 @@
 from datetime import datetime, date
-from sqlalchemy import Column, BigInteger, String, Date, SmallInteger, Text, DateTime, ForeignKey
+from sqlalchemy import Integer, Column, BigInteger, String, Date, SmallInteger, Text, DateTime, ForeignKey
 from app.core.database import Base
 
 class Acceptance(Base):
     __tablename__ = "acceptance"
-    id = Column(BigInteger, primary_key=True, autoincrement=True)
+    id = Column(Integer, primary_key=True, autoincrement=True)
     project_id = Column(BigInteger, ForeignKey("projects.id"), nullable=False)
     title = Column(String(200), nullable=False)
     acceptance_date = Column(Date)
