+++ backend/app/models/ai_work_summary.py
@@ -1,5 +1,5 @@
 from datetime import datetime
-from sqlalchemy import Column, BigInteger, String, Text, DateTime
+from sqlalchemy import Integer, Column, BigInteger, String, Text, DateTime
 from app.core.database import Base
 
 REPORT_TYPES = {"daily": "日报", "weekly": "周报", "monthly": "月报"}
@@ -7,7 +7,7 @@ REPORT_TYPES = {"daily": "日报", "weekly": "周报", "monthly": "月报"}
 
 class AiWorkSummary(Base):
     __tablename__ = "ai_work_summary"
-    id = Column(BigInteger, primary_key=True, autoincrement=True)
+    id = Column(Integer, primary_key=True, autoincrement=True)
     report_type = Column(String(20), nullable=False)
     title = Column(String(200), nullable=False)
     content = Column(Text, nullable=False)
