+++ backend/app/models/decision_chain.py
@@ -1,10 +1,10 @@
 from datetime import datetime
-from sqlalchemy import Column, BigInteger, String, SmallInteger, Text, DateTime, ForeignKey
+from sqlalchemy import Integer, Column, BigInteger, String, SmallInteger, Text, DateTime, ForeignKey
 from app.core.database import Base
 
 class DecisionChain(Base):
     __tablename__ = "decision_chain"
-    id = Column(BigInteger, primary_key=True, autoincrement=True)
+    id = Column(Integer, primary_key=True, autoincrement=True)
     project_id = Column(BigInteger, ForeignKey("projects.id"), nullable=False)
     name = Column(String(100), nullable=False)
     role_type = Column(String(50), nullable=False)
